refactor(workbench): move files to their designated folders
