+++ projects/scion/workbench/src/lib/view/view-activation-instant-provider.service.ts
