+++ projects/scion/workbench/src/lib/view/view-move-handler.service.ts
@@ -10,10 +10,10 @@ import { ɵWorkbenchService } from '../ɵworkbench.service';
 import { WorkbenchRouter } from '../routing/workbench-router.service';
 
 /**
- * Subscribes to view move requests for moving views in the {@link PartsLayout} when the user arranges views via drag and drop.
+ * Subscribes to view move requests for moving views in the {@link PartsLayout} when the user arranges views via drag and drop or view context menu.
  */
 @Injectable({providedIn: 'root'})
-export class ViewDropHandler implements OnDestroy {
+export class ViewMoveHandler implements OnDestroy {
 
   private _destroy$ = new Subject<void>();
